Update Rust toolchain to 1.77.2, pass or suppress lints, use new features
core::mem::offset_of! was stabilized in Rust 1.77.0, but there is no equivalent to memoffset::span_of!.
The ranges are easy to write manually, but that would risk introducing bugs unless we added more tests.
It may be better to specify offsets explicitly instead of having a struct just to calculate them.
None of that belongs in a toolchain update.

None of the slice methods stabilized in Rust 1.77.0 appear to be useful.
The shuffling crate needs [T]::as_chunks_mut and [T]::as_rchunks_mut, which are still unstable.