Remove incorrect SszRead impl for types::combined::AttesterSlashing
It appears to be impossible to implement it correctly.
AttesterSlashing does not contain any field that can be used to determine its phase.
The attestations inside AttesterSlashing may be from a different phase.

Rewrite fork_choice_control::spec_tests accordingly.
Add new methods to types::config::Config to support it.

AttesterSlashing now has to be tested the same way as the light client containers.
AttesterSlashing::phase is no longer needed.