Return HTTP status code 415 from endpoints that expect JSON (#26)
axum::Json already does that, but our EthJson and Error were replacing its status codes with 400.
The solution was to store JsonRejection directly in Error and delegate to it in Error::status_code.
Doing so also makes endpoints return status codes 413 and 422 when appropriate.

The duplicated rejection messages came up again.
We were able to determine the cause and come up with a better workaround this time.
The issue should be fixed for query strings too.