Fix remaining compilation errors
eth2_libp2p needed changes as expected.

The changes on this branch are making some tests fail.
They are failing in 5 crates as opposed to 1 on electra-alpha3-containers.
That is most likely due to changed definitions of hash_tree_root and BlobSidecar.
There is a branch of consensus-spec-tests for EIP-7688, but it may be based on a different version.

The oversight with proof lengths should be fixed.
The cause of the crash while attesting is still unknown.
It may be specific to Kurtosis.
The changes made on this branch should not have affected anything until Electra (now Deneb).
The built-in-validator test case only fails in slot 40 (Electra), which is unchanged from electra-alpha3-containers.